forms: phase 1.2d — extract TextBox natives + textconv
Continue moving per-kind natives. This commit pulls the EDIT-control
cluster out of forms_module.c and adds a small UTF-8 / wide-char
helper TU that the TextBox natives (and several other future TUs)
need.

  src/core/textconv.{c,h}
    - utf8_to_wide()       — alloc + MultiByteToWideChar
    - wide_to_utf8()       — alloc + WideCharToMultiByte
    - utf8_into_wide_buf() — fixed-buffer variant for command queue
                              text payloads.
    Win32-only on the inside; non-Windows builds get NULL/no-op stubs
    so the .so still links even though no caller exercises them.

  src/controls/ctl_textbox.{c,h}
    - native_set_multiline      (ES_MULTILINE / ES_AUTOVSCROLL toggle)
    - native_set_readonly       (EM_SETREADONLY)
    - native_set_placeholder    (EM_SETCUEBANNER)
    - native_set_password_char  (EM_SETPASSWORDCHAR)
    - native_set_max_length     (EM_SETLIMITTEXT)
    - native_set_text_alignment (ES_LEFT/CENTER/RIGHT toggle, also
                                 reused by Label/LinkLabel via
                                 SS_LEFT/CENTER/RIGHT)
    - native_select_all         (EM_SETSEL 0..-1)
    - native_append_text        (EM_SETSEL + EM_REPLACESEL)
    - native_clear_text         (SetWindowTextW for TextBox,
                                 CB_RESETCONTENT for ComboBox,
                                 LB_RESETCONTENT for ListBox)
    - edit_toggle_style helper  — applies style flags to a Win32
                                  control with SWP_FRAMECHANGED reflow.

forms_module.c now #includes textconv.h and ctl_textbox.h in place
of the in-place definitions. Every Phase 1.2 TU uses the same
test-build / Win32 gating pattern: bodies wrapped in
#ifndef FORMS_MODULE_TEST_BUILD, Win32-specific code wrapped in
#if defined(CANDO_PLATFORM_WINDOWS) || _WIN32 || _WIN64.

Main TU shrinks from 4571 → 4346 lines (cumulative -891 lines from
the 5237-line baseline).

All 100 C unit tests still pass.
