+++ modules/forms/Makefile
@@ -35,20 +35,23 @@ CFLAGS     = -std=c11 -Wall -Wextra -Wpedantic -fPIC $(INCLUDES)
 # identically for the .so / .dll / test builds.
 CORE_C     = src/core/color.c        src/core/geom.c   src/core/events.c \
              src/core/slots.c        src/core/dispatch.c \
+             src/core/textconv.c \
              src/controls/ctl_common.c \
              src/controls/ctl_checkbox.c \
              src/controls/ctl_progress.c \
              src/controls/ctl_trackbar.c \
-             src/controls/ctl_numeric.c
+             src/controls/ctl_numeric.c \
+             src/controls/ctl_textbox.c
 CORE_H     = src/core/color.h        src/core/geom.h   src/core/events.h \
              src/core/sync.h         src/core/cando_compat.h \
              src/core/slots.h        src/core/manager.h \
-             src/core/dispatch.h \
+             src/core/dispatch.h     src/core/textconv.h \
              src/controls/ctl_common.h \
              src/controls/ctl_checkbox.h \
              src/controls/ctl_progress.h \
              src/controls/ctl_trackbar.h \
-             src/controls/ctl_numeric.h
+             src/controls/ctl_numeric.h \
+             src/controls/ctl_textbox.h
 
 ifeq ($(OS),Windows_NT)
     SHARED      = forms.dll
